Split telepath blocks into separate files
